DRY up eslint config
