+++ packages/form-section/package.json
@@ -44,7 +44,7 @@
     "madge": "^8.0.0",
     "prettier": "^3.4.2",
     "tslib": "^2.8.1",
-    "typescript": "~5.7.2"
+    "typescript": "^5.7.2"
   },
   "eslintConfig": {
     "extends": [
